fix: correct seppuku type for seconds

+++ apps/api-gql/internal/gql/resolvers/games.resolver.service.go
@@ -373,7 +373,7 @@ func (r *mutationResolver) gamesUpdateSeppuku(
 	}
 
 	if opts.TimeoutSeconds.IsSet() {
-		entity.TimeoutSeconds = int8(*opts.TimeoutSeconds.Value())
+		entity.TimeoutSeconds = *opts.TimeoutSeconds.Value()
 	}
 
 	if opts.TimeoutModerators.IsSet() {
